src/Bots/client/requirements.txt
pika==1.1.0
requests==2.24.0
urllib3==1.25.10
python-dotenv==0.15.0
